requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
amqp==5.0.6
    # via kombu
asgiref==3.4.1
    # via django
attrs==21.2.0
    # via
    #   jsonschema
    #   pytest
billiard==3.6.4.0
    # via celery
cachecontrol==0.12.10
    # via firebase-admin
cachetools==4.2.4
    # via google-auth
celery==5.1.2
    # via -r requirements.in
certifi==2021.10.8
    # via requests
charset-normalizer==2.0.7
    # via requests
click==7.1.2
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
coverage[toml]==6.0.2
    # via pytest-cov
django==3.2.8
    # via
    #   -r requirements.in
    #   django-admin-sortable2
    #   django-cors-headers
    #   django-filter
    #   django-model-utils
    #   django-redis
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-spectacular
django-admin-sortable2==1.0.2
    # via -r requirements.in
django-cors-headers==3.10.0
    # via -r requirements.in
django-daterangefilter==1.0.0
    # via -r requirements.in
django-filter==21.1
    # via -r requirements.in
django-model-utils==4.2.0
    # via -r requirements.in
django-redis==5.0.0
    # via -r requirements.in
djangorestframework==3.12.4
    # via
    #   -r requirements.in
    #   djangorestframework-gis
    #   djangorestframework-simplejwt
    #   drf-spectacular
djangorestframework-gis==0.17
    # via -r requirements.in
djangorestframework-simplejwt==5.0.0
    # via -r requirements.in
drf-spectacular==0.20.2
    # via -r requirements.in
environ==1.0
    # via -r requirements.in
factory-boy==3.2.0
    # via
    #   -r requirements.in
    #   pytest-factoryboy
faker==9.3.1
    # via
    #   -r requirements.in
    #   factory-boy
firebase-admin==5.1.0
    # via -r requirements.in
flake8==4.0.1
    # via -r requirements.in
google-api-core[grpc]==2.2.2
    # via
    #   firebase-admin
    #   google-api-python-client
    #   google-cloud-core
    #   google-cloud-firestore
    #   google-cloud-storage
google-api-python-client==2.31.0
    # via firebase-admin
google-auth==2.3.3
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-cloud-core
    #   google-cloud-storage
google-auth-httplib2==0.1.0
    # via google-api-python-client
google-cloud-core==2.2.1
    # via
    #   google-cloud-firestore
    #   google-cloud-storage
google-cloud-firestore==2.3.4
    # via firebase-admin
google-cloud-storage==1.43.0
    # via firebase-admin
google-crc32c==1.3.0
    # via google-resumable-media
google-resumable-media==2.1.0
    # via google-cloud-storage
googleapis-common-protos==1.53.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.42.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.42.0
    # via google-api-core
gunicorn==20.1.0
    # via -r requirements.in
httplib2==0.20.2
    # via
    #   google-api-python-client
    #   google-auth-httplib2
idna==3.3
    # via requests
inflection==0.5.1
    # via
    #   drf-spectacular
    #   pytest-factoryboy
iniconfig==1.1.1
    # via pytest
isort==5.10.0
    # via -r requirements.in
jsonschema==4.1.0
    # via drf-spectacular
kombu==5.1.0
    # via celery
mccabe==0.6.1
    # via flake8
msgpack==1.0.2
    # via cachecontrol
mypy==0.910
    # via -r requirements.in
mypy-extensions==0.4.3
    # via mypy
packaging==21.0
    # via
    #   google-cloud-firestore
    #   pytest
pillow==8.4.0
    # via -r requirements.in
pluggy==1.0.0
    # via pytest
prompt-toolkit==3.0.20
    # via click-repl
proto-plus==1.19.8
    # via google-cloud-firestore
protobuf==3.19.1
    # via
    #   google-api-core
    #   google-cloud-storage
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
psycopg2==2.9.1
    # via -r requirements.in
py==1.10.0
    # via pytest
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pycodestyle==2.8.0
    # via flake8
pyflakes==2.4.0
    # via flake8
pyjwt==2.2.0
    # via
    #   djangorestframework-simplejwt
    #   twilio
pyparsing==2.4.7
    # via
    #   httplib2
    #   packaging
pyrsistent==0.18.0
    # via jsonschema
pytest==6.2.5
    # via
    #   -r requirements.in
    #   pytest-cov
    #   pytest-django
    #   pytest-factoryboy
pytest-cov==3.0.0
    # via -r requirements.in
pytest-django==4.4.0
    # via -r requirements.in
pytest-factoryboy==2.1.0
    # via -r requirements.in
python-dateutil==2.8.2
    # via faker
pytz==2021.3
    # via
    #   celery
    #   django
    #   twilio
pyyaml==6.0
    # via drf-spectacular
redis==3.5.3
    # via django-redis
requests==2.26.0
    # via
    #   -r requirements.in
    #   cachecontrol
    #   google-api-core
    #   google-cloud-storage
    #   twilio
rsa==4.7.2
    # via google-auth
six==1.16.0
    # via
    #   click-repl
    #   google-auth
    #   google-auth-httplib2
    #   google-cloud-storage
    #   grpcio
    #   python-dateutil
sqlparse==0.4.2
    # via django
supervisor==4.2.2
    # via -r requirements.in
text-unidecode==1.3
    # via faker
toml==0.10.2
    # via
    #   mypy
    #   pytest
tomli==1.2.1
    # via coverage
twilio==7.1.0
    # via -r requirements.in
typing-extensions==3.10.0.2
    # via mypy
ua-parser==0.10.0
    # via -r requirements.in
uritemplate==4.1.1
    # via
    #   drf-spectacular
    #   google-api-python-client
urllib3==1.26.7
    # via requests
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.5
    # via prompt-toolkit

# The following packages are considered to be unsafe in a requirements file:
# setuptools
